OPHAKTKEH-283 Reorganise authorisation listing
