+++ src/main/reactjs/src/components/clerkTranslator/overview/AuthorisationDetails.tsx
@@ -1,121 +1,12 @@
-import {
-  Add as AddIcon,
-  DeleteOutline as DeleteIcon,
-  InfoOutlined as InfoIcon,
-} from '@mui/icons-material';
-import {
-  Switch,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableRow,
-  Tooltip,
-} from '@mui/material';
+import { Add as AddIcon } from '@mui/icons-material';
 
+import { AuthorisationListing } from 'components/clerkTranslator/overview/AuthorisationListing';
 import { CustomButton } from 'components/elements/CustomButton';
-import { CustomIconButton } from 'components/elements/CustomIconButton';
 import { H3, Text } from 'components/elements/Text';
-import {
-  useAppTranslation,
-  useCommonTranslation,
-  useKoodistoLanguagesTranslation,
-} from 'configs/i18n';
+import { useAppTranslation } from 'configs/i18n';
 import { useAppSelector } from 'configs/redux';
 import { Color, Variant } from 'enums/app';
-import { AuthorisationBasisEnum } from 'enums/clerkTranslator';
-import { Authorisation } from 'interfaces/authorisation';
 import { clerkTranslatorOverviewSelector } from 'redux/selectors/clerkTranslatorOverview';
-import { AuthorisationUtils } from 'utils/authorisation';
-import { DateUtils } from 'utils/date';
-
-const AuthorisationsListing = ({
-  authorisations,
-}: {
-  authorisations: Array<Authorisation>;
-}) => {
-  const translateLanguage = useKoodistoLanguagesTranslation();
-  const translateCommon = useCommonTranslation();
-  const { t } = useAppTranslation({
-    keyPrefix: 'akt.component.clerkTranslatorOverview.authorisations',
-  });
-  const currentDate = new Date();
-
-  return (
-    <Table
-      size="small"
-      className="clerk-translator-details__authorisations-table"
-    >
-      <TableHead>
-        <TableRow>
-          <TableCell>{t('fields.languagePair')}</TableCell>
-          <TableCell>{t('fields.basis')}</TableCell>
-          <TableCell>{t('fields.startDate')}</TableCell>
-          <TableCell>{t('fields.endDate')}</TableCell>
-          <TableCell>{t('fields.isEffective')}</TableCell>
-          <TableCell>{t('fields.permissionToPublish')}</TableCell>
-          <TableCell>{t('fields.diaryNumber')}</TableCell>
-          <TableCell>{translateCommon('delete')}</TableCell>
-        </TableRow>
-      </TableHead>
-      <TableBody>
-        {authorisations.map((a, i) => (
-          <TableRow
-            key={i}
-            data-testid={`authorisations-table__id-${a.id}-row`}
-          >
-            <TableCell>
-              <Text>
-                {`${translateLanguage(a.languagePair.from)}
-             - ${translateLanguage(a.languagePair.to)}`}
-              </Text>
-            </TableCell>
-            <TableCell>
-              <div className="columns gapped-xs">
-                <Text>{a.basis}</Text>
-                {a.basis === AuthorisationBasisEnum.AUT && (
-                  <Tooltip
-                    title={`${t(
-                      'fields.autDate'
-                    )}: ${DateUtils.formatOptionalDate(a.autDate)}`}
-                    arrow
-                    placement="bottom"
-                  >
-                    <CustomIconButton>
-                      <InfoIcon color={Color.Secondary} />
-                    </CustomIconButton>
-                  </Tooltip>
-                )}
-              </div>
-            </TableCell>
-            <TableCell>
-              <Text>{DateUtils.formatOptionalDate(a.termBeginDate)}</Text>
-            </TableCell>
-            <TableCell>
-              <Text>{DateUtils.formatOptionalDate(a.termEndDate)}</Text>
-            </TableCell>
-            <TableCell>
-              <Text>
-                {AuthorisationUtils.isAuthorisationEffective(a, currentDate)
-                  ? translateCommon('yes')
-                  : translateCommon('no')}
-              </Text>
-            </TableCell>
-            <TableCell>
-              <Switch checked={a.permissionToPublish} color={Color.Secondary} />
-            </TableCell>
-            <TableCell>
-              <Text>{a.diaryNumber}</Text>
-            </TableCell>
-            <TableCell className="centered">
-              <DeleteIcon className="color-red-500" />
-            </TableCell>
-          </TableRow>
-        ))}
-      </TableBody>
-    </Table>
-  );
-};
 
 export const AuthorisationDetails = () => {
   const { t } = useAppTranslation({
@@ -139,11 +30,14 @@ export const AuthorisationDetails = () => {
           {t('buttons.add')}
         </CustomButton>
       </div>
-      <AuthorisationsListing
-        authorisations={
-          selectedTranslator ? selectedTranslator.authorisations : []
-        }
-      />
+
+      {selectedTranslator?.authorisations.length ? (
+        <AuthorisationListing
+          authorisations={selectedTranslator.authorisations}
+        />
+      ) : (
+        <Text>{t('noAuthorisations')}</Text>
+      )}
     </>
   );
 };
